+++ Service/src/Dt.Core/Api/Base/DataAccess.cs
@@ -121,7 +121,7 @@ namespace Dt.Core
         /// </summary>
         /// <param name="p_seqName">序列名称，不可为空</param>
         /// <returns>新序列值</returns>
-        public Task<int> NewSeq(string p_seqName)
+        public Task<long> NewSeq(string p_seqName)
         {
             return _da.NewSeq(p_seqName);
         }
