+++ Service/src/Dt.Core/Base/DataAccess/Access/SqlServerAccess.cs
@@ -43,11 +43,11 @@ namespace Dt.Core
             return $"{sql} offset {p_starRow} rows fetch next {p_pageSize} rows only";
         }
 
-        public override Task<int> NewSeq(string p_seqName)
+        public override Task<long> NewSeq(string p_seqName)
         {
             if (!string.IsNullOrEmpty(p_seqName))
-                return GetScalar<int>($"SELECT NEXT VALUE FOR [{p_seqName}]");
-            return Task.FromResult(0);
+                return GetScalar<long>($"SELECT NEXT VALUE FOR [{p_seqName}]");
+            return Task.FromResult(0L);
         }
         #endregion
